# CPLAT-10490 - Add a config that allows collapsing directories with no excludes. 🏋️‍♀️