# Problem
We have a repository that has over 3000 files to format.  Mac OS, at least, had a hardcoded kernel setting that causes that argument list to be too long.  Because of that, we aren't currently able to format in one pass.

# Solution
Allow the format tool to collapse directories that have no excludes.  In our repo, this cut the paths from 3012 to 705.  The default still expands directories, so consumers will not experience a change in behavior unless they opt-in.

# Testing
- [ ] Pull this version of dart_dev into your repo.  Run dart_dev format, and verify it formats the same files as before, but using many less input paths.


## Example run with timings

### Without Collapse

```
$ doc_plat_client git:(PP-3396_blackline_review_cards_rebase) ✗ time ddev format
Matching exclude glob dirs:
[INFO] excluding 728 files
[INFO] Running subprocess:
pub run over_react_format --line-length=120 <2835 paths>
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Formatting 2835 files...
    2835 files unchanged
Success! All files are already formatted correctly.
pub run dart_dev format  79.78s user 11.79s system 334% cpu 27.380 total
```

### With Collapse

```
$doc_plat_client git:(PP-3396_blackline_review_cards_rebase) ✗ time ddev format
Matching exclude glob dirs:
Glob didn't match any paths
[INFO] excluding 728 files
[INFO] Running subprocess:
pub run over_react_format --line-length=120 <707 paths>
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Formatting 2835 files...
    2835 files unchanged
Success! All files are already formatted correctly.
pub run dart_dev format  77.91s user 11.95s system 342% cpu 26.256 total
```