@tomconnell-wf 2020-04-15
Rebased from https://github.com/Workiva/dart_dev/pull/340