@tomconnell-wf 2020-04-15
I'll need to make a tiny PR to over_react_format to have it pass through a collapseDirectories option.