@tomconnell-wf 2020-04-15
With things provisionally wired up through over_react_format:
![image](https://user-images.githubusercontent.com/8027264/79391518-f1318900-7f2e-11ea-8af2-7fba4beaea40.png)

Notice the number of paths passed in versus the number of files formatted.  They used to be lockstep.