@tomconnell-wf 2020-05-07
@evanweible-wf I think this addresses the problem.  It uses Glob.listSync, which gets all the FileSystemEntities to be excluded.  I think make a set of the directories that contain those files.  Then, when going through the files to format, if a directory contains any of the excluded directory set, it keeps descending and does not collapse.  The default behavior should remain the same as before this PR.