@tomconnell-wf 2020-05-07
I tested on a wider array of repos that I knew were using dart_dev ^3.0.0 and saw the same files formatted with and without collapse.