@tomconnell-wf 2020-05-08
I tested doc_plat_client, graph_app, wdesk_sdk, audit, w_outline, app_intelligence, and over react with collapse behavior and without.  It wound up excluding the same number of files and formatting the same number of files.  I checked the actual list of excluded files in a couple of the repos, and they were the same.

Of interest was w_comments.  It has no excludes, so doesn't benefit from the collapse behavior, but since there isn't a performance implication, I don't want to look into it.